# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Company Ship API",
        "description": "API quản lý vận tải - công ty, nhân sự, xe, chuyến đi, lương",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/drivers": {
            "get": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Danh sách tài xế",
                "operationId": "a75c8627a59b2479de780c37b0c06e03",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Tìm theo license_no",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "employee_id",
                        "in": "query",
                        "description": "Lọc theo nhân viên",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "available_status",
                        "in": "query",
                        "description": "Lọc theo trạng thái",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sắp xếp",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Số bản ghi/trang",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    }
                }
            },
            "post": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Tạo tài xế mới",
                "operationId": "829ed7a657b54c9727b94e7b814c2497",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "employee_id",
                                    "license_no",
                                    "license_class"
                                ],
                                "properties": {
                                    "employee_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "license_no": {
                                        "type": "string",
                                        "example": "B123456"
                                    },
                                    "license_class": {
                                        "type": "string",
                                        "example": "B2"
                                    },
                                    "expired_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "available_status": {
                                        "type": "string",
                                        "enum": [
                                            "available",
                                            "on_trip",
                                            "off"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo thành công"
                    },
                    "422": {
                        "description": "Validation lỗi"
                    }
                }
            }
        },
        "/api/drivers/{id}": {
            "get": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Chi tiết tài xế",
                "operationId": "b8cec5341d14f654472f26d4c0771b61",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            },
            "put": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Cập nhật tài xế",
                "operationId": "5b5192462e5b1b239c63d12bda639848",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "employee_id": {
                                        "type": "integer"
                                    },
                                    "license_no": {
                                        "type": "string"
                                    },
                                    "license_class": {
                                        "type": "string"
                                    },
                                    "expired_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "available_status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    },
                    "422": {
                        "description": "Validation lỗi"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Xóa tài xế",
                "operationId": "8351d443e6a33b2c835fb86af7782d12",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Dùng token trả về từ POST /api/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}